upload images

+++ src/app/components/users/users.component.ts
@@ -3,6 +3,7 @@ import { UserService } from './../../services/user.service';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ToastrService } from 'ngx-toastr';
 import { Statistics } from './../../interfaces/statistics';
+import { NgbActiveModal, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-users',
@@ -14,7 +15,8 @@ export class UsersComponent implements OnInit, OnDestroy {
   subscribers: any[] = [];
   statistics: Statistics = {} as Statistics;
 
-  constructor(private userService: UserService, private toastr: ToastrService) {
+  constructor(private userService: UserService,
+    private toastr: ToastrService) {
     this.subscribers.push(
       this.userService.getAllUsers().subscribe((users) => {
         return (this.users = users);
@@ -29,12 +31,19 @@ export class UsersComponent implements OnInit, OnDestroy {
         .subscribe((statistics) => (this.statistics = statistics))
     );
   }
+
   removeUser(userId: string | number) {
-    this.subscribers.push(
-      this.userService.removeUser(userId).subscribe((message) => {
-        this.toastr.success(`${message}`);
-      })
-    );
+    if (window.confirm('Are sure you want to delete this user ?')) {
+      this.subscribers.push(
+        this.userService.removeUser(userId).subscribe((message) => {
+          this.toastr.success(`${message}`);
+        })
+      )
+      setTimeout(() => {
+        window.location.reload()
+      }, 1000);
+    }
+
   }
 
   ngOnDestroy(): void {
